Move teams tests to separate file
